fitness-agent-full-v2/requirements.txt
python-dotenv>=1.0.1
requests>=2.32.3
pydantic>=2.7.0
PyYAML>=6.0.1
jsonschema>=4.22.0
